extract MemoryRecorder & LogHandler
