+++ Assets/Scripts/Runtime/MemoryRecorder.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 03504356ffec74709b0e94d6591fe79a
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
