+++ Assets/Scripts/Runtime/RtXrQuestApplication.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: d2d33d447ee9d4653ac750dba81a1097
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
